--- a/templates/query_default.json.tmpl
+++ b/templates/query_default.json.tmpl
@@ -1,4 +1,4 @@
 {
     "top_k": 10,
-    "vector": {{ vector 768 }}
+    "rank_by": ["vector", "ANN", {{ vector 768 }}]
 }