Now that we have an official Go SDK [0], use it.

This neatly removes a lot of handwritten API call code. We don't get to use the full power of the SDK because this tool renders its own JSON for query and write requests, but that's its own cool validation of the escape hatches that Stainless provides.

Perf looks equivalent to before, in this one relatively unscientific test.

Old (no SDK):

```
$ ./tpuf-benchmark -api-key tpuf_... -endpoint https://gcp-us-central1.turbopuffer.com -namespace-count 100 -namespace-combined-size 1000000 -upserts-per-sec 0 -queries-per-sec 1000     -query-template templates/query_default.json.tmpl -benchmark-duration 60s

clearing namespaces 100% |████████████████████████████████████| (100/100, 113 it/s)        
upserting documents 100% |█████████████████████████| (1000000/1000000, 24135 it/s)         
2025/06/05 19:52:15 waiting for 100 namespace(s) to be indexed before starting benchmark
2025/06/05 19:52:15 all namespaces have been (reasonably) indexed
2025/06/05 19:52:15 starting benchmark, running for 1m0s
2025/06/05 19:52:15 executing queries at 1000.00 QPS (uniform distribution)

Report for the last 10s:
hot_queries:
  count: 9700
  latencies: min=11ms, p10=14ms, p25=16ms, p50=18ms, p75=21ms, p90=25ms, p95=28ms, p99=47ms, max=208ms

Report for the last 10s:
hot_queries:
  count: 9680
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=20ms, p90=23ms, p95=26ms, p99=44ms, max=89ms

Report for the last 10s:
hot_queries:
  count: 9698
  latencies: min=11ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=22ms, p95=25ms, p99=40ms, max=96ms

Report for the last 10s:
hot_queries:
  count: 9711
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=22ms, p95=25ms, p99=38ms, max=89ms

Report for the last 10s:
hot_queries:
  count: 9691
  latencies: min=11ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=22ms, p95=25ms, p99=40ms, max=85ms
2025/06/05 19:53:15 benchmark duration of 1m0s has elapsed, shutting down

Cumulative report for the entire benchmark:
hot_queries:
  count: 58140
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=23ms, p95=26ms, p99=42ms, max=208ms
2025/06/05 19:53:15 results written to: benchmark-results-1749153088
```

New (with SDK):

```
clearing namespaces 100% |████████████████████████████████████| (100/100, 169 it/s)        
upserting documents 100% |█████████████████████████| (1000000/1000000, 27406 it/s)         
2025/06/05 20:31:02 waiting for 100 namespace(s) to be indexed before starting benchmark
2025/06/05 20:31:02 all namespaces have been (reasonably) indexed
2025/06/05 20:31:02 starting benchmark, running for 1m0s
2025/06/05 20:31:02 executing queries at 1000.00 QPS (uniform distribution)

Report for the last 10s:
hot_queries:
  count: 9681
  latencies: min=11ms, p10=14ms, p25=16ms, p50=18ms, p75=21ms, p90=25ms, p95=29ms, p99=45ms, max=86ms

Report for the last 10s:
hot_queries:
  count: 9707
  latencies: min=10ms, p10=14ms, p25=16ms, p50=17ms, p75=20ms, p90=23ms, p95=27ms, p99=46ms, max=124ms

Report for the last 10s:
hot_queries:
  count: 9675
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=23ms, p95=27ms, p99=44ms, max=78ms

Report for the last 10s:
hot_queries:
  count: 9699
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=22ms, p95=25ms, p99=42ms, max=69ms

Report for the last 10s:
hot_queries:
  count: 9670
  latencies: min=11ms, p10=14ms, p25=15ms, p50=17ms, p75=19ms, p90=22ms, p95=25ms, p99=40ms, max=71ms
2025/06/05 20:32:02 benchmark duration of 1m0s has elapsed, shutting down

Cumulative report for the entire benchmark:
hot_queries:
  count: 58099
  latencies: min=10ms, p10=14ms, p25=15ms, p50=17ms, p75=20ms, p90=23ms, p95=26ms, p99=44ms, max=124ms
2025/06/05 20:32:02 results written to: benchmark-results-1749155419
```

[0]: https://github.com/turbopuffer/turbopuffer-go